services/order service/requirements.txt
django
djangorestframework
psycopg2-binary
requests
pika